requirements.txt
#
# This file is autogenerated by pip-compile with python 3.10
# To update, run:
#
#    pip-compile --resolver=backtracking
#
asgiref==3.5.2
    # via django
cffi==1.15.1
    # via cryptography
cryptography==38.0.3
    # via -r requirements.in
django==4.1.3
    # via
    #   -r requirements.in
    #   django-phone-field
    #   djangorestframework
    #   djangorestframework-simplejwt
django-phone-field==1.8.1
    # via -r requirements.in
djangorestframework==3.14.0
    # via
    #   -r requirements.in
    #   djangorestframework-simplejwt
djangorestframework-simplejwt==5.2.2
    # via -r requirements.in
psycopg2==2.9.5
    # via -r requirements.in
pycparser==2.21
    # via cffi
pyjwt==2.6.0
    # via djangorestframework-simplejwt
pytz==2022.6
    # via djangorestframework
sqlparse==0.4.3
    # via django
